pyproject.toml
[tool.poetry]
name = "quantitative-research-platform"
version = "1.0.0"
description = "Comprehensive Quantitative Trading Research Platform - Advanced framework for quant analysis, backtesting, and ML model development"
authors = ["Quantitative Research Team <quant@hivecapital.ai>"]
readme = "README.md"
packages = [{include = "quant_research"}, {include = "agents"}, {include = "strategies"}, {include = "data"}, {include = "config"}]

[tool.poetry.dependencies]
python = "^3.11"
# Configuration and environment
pydantic = "^2.7.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
# Database & Storage
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
asyncpg = "^0.29.0"
alembic = "^1.13.0"
redis = "^5.0.0"
# High-Performance Data Processing
polars = "^1.33.0"
numba = "^0.61.2"
vectorbt = "^0.28.1"
# LangGraph and AI
langgraph = "^0.0.40"
langchain = "^0.1.0"
langchain-core = "^0.1.0"
# Sentiment Analysis and NLP
transformers = "^4.36.0"
torch = "^2.1.0"
google-generativeai = "^0.3.0"
openai = "^1.3.0"
feedparser = "^6.0.10"
beautifulsoup4 = "^4.12.0"
requests = "^2.31.0"
newspaper3k = "^0.2.8"
textblob = "^0.17.1"
# Market Data APIs
alpaca-trade-api = "^3.2.0"
alpaca-py = "^0.42.0"
polygon-api-client = "^1.12.0"
yfinance = "^0.2.65"
ccxt = "^4.5.2"
alpha-vantage = "^3.0.0"
fredapi = "^0.5.2"
finvizfinance = "^1.1.1"
# Data processing
pandas = "^2.3.2"
numpy = "^2.2.6"
scipy = "^1.16.1"
statsmodels = "^0.14.5"
arch = "^7.2.0"
# Visualization
matplotlib = "^3.10.5"
mplfinance = "^0.12.10"
seaborn = "^0.13.2"
plotly = "^6.3.0"
dash = "^3.2.0"
# Web Framework
flask = "^3.0.0"
# GUI
tk = "^0.1.0"
# Utilities
click = "^8.1.0"
rich = "^13.7.0"
typer = "^0.9.0"
# Security
cryptography = "^41.0.0"
# Logging
structlog = "^23.2.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
# Code quality
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
pre-commit = "^3.6.0"
# Development tools
jupyter = "^1.0.0"
ipython = "^8.17.0"
# Documentation
mkdocs = "^1.5.0"
mkdocs-material = "^9.4.0"

[tool.poetry.group.quant.dependencies]
# Quantitative Finance Libraries
QuantLib = "^1.39"
QuantLib-Python = "^1.18"
FinancePy = "^1.0.1"
py-vollib = "^1.0.1"
Riskfolio-Lib = "^7.0.1"
empyrical-reloaded = "^0.5.12"
QuantStats = "^0.0.76"
ta = "^0.11.0"
TA-Lib = "^0.6.6"
exchange-calendars = "^4.11.1"
pandas-market-calendars = "^5.1.1"

[tool.poetry.group.backtest.dependencies]
# Backtesting Frameworks
backtrader = "^1.9.78.123"
vectorbt = "^0.28.1"
fastquant = "^0.1.8.1"
alphalens-reloaded = "^0.4.6"

[tool.poetry.group.ml.dependencies]
# Machine Learning & AI
scikit-learn = "^1.7.1"
xgboost = "^3.0.4"
lightgbm = "^4.6.0"
stable-baselines3 = "^2.7.0"
gymnasium = "^1.2.0"
FinRL = "^0.3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["agents", "strategies", "data", "config"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "alpaca.*",
    "polygon.*",
    "ta.*",
    "backtrader.*",
    "vectorbt.*",
    "kafka.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests", 
    "backtest: Backtesting tests",
    "slow: Slow running tests",
    "live: Tests requiring live market data"
]

[tool.coverage.run]
source = ["agents", "strategies", "data", "config"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]